feat(ios): add image evaluation from file path
